go-multimap: Added implementation for setmultimap
This diff is a rather large refactor of the previous implementation. The
previous multimap implementation only uses a slice based multimap, and
when there are duplicates, it is added to the multimap as a separate
value. This commit adds an implementation of a multimap which does not allow duplicates.

The implementation only uses the go standard library. This addresses #2.